requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
youtube-transcript-api
langchain
langchain-community
langchain-core
langchain-google-genai
faiss-cpu